#30 Splitted home component to module
