+++ frontend/src/app/core/auth/login.guard.ts
@@ -1,7 +1,7 @@
 import { CanActivateFn, Router } from '@angular/router';
 import { inject } from "@angular/core";
 import { UserService } from "./user.service";
-import { catchError, map, of, take } from "rxjs";
+import { catchError, map, of } from "rxjs";
 
 export const LoginGuard: CanActivateFn = (route, state) => {
   const userService = inject(UserService);
