+++ frontend/src/app/modules/core/core.module.ts
@@ -5,14 +5,12 @@ import {CoreRoutingModule} from './core-routing.module';
 import {CoreComponent} from './core.component';
 import {GeneralModule} from "../../core/general.module";
 import {NavComponent} from "../../core/nav/nav.component";
-import {HomeComponent} from "../../core/home/home.component";
 
 
 @NgModule({
   declarations: [
     CoreComponent,
     NavComponent,
-    HomeComponent
   ],
   imports: [
     GeneralModule,
