+++ frontend/src/app/modules/core/home/home.component.scss
