+++ frontend/src/app/modules/core/home/home.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { HomeRoutingModule } from './home-routing.module';
+import {SharedModule} from "../../../shared/shared.module";
+import {HomeComponent} from "./home.component";
+
+
+@NgModule({
+  declarations: [
+    HomeComponent
+  ],
+  imports: [
+    CommonModule,
+    SharedModule,
+    HomeRoutingModule
+  ]
+})
+export class HomeModule { }
